更新源

+++ packages/af-webpack/src/build.js
@@ -22,10 +22,13 @@ export default function build(opts = {}) {
     'webpackConfig should be plain object or array.',
   );
 
-  // 清理 output path
   const outputPath = getOutputPath(webpackConfig);
-  debug(`Clean output path ${outputPath.replace(`${cwd}/`, '')}`);
-  rimraf.sync(outputPath);
+
+  // 清理 output path
+  if (process.env.CLEAR_OUTPUT !== 'none') {
+    debug(`Clean output path ${outputPath.replace(`${cwd}/`, '')}`);
+    rimraf.sync(outputPath);
+  }
 
   debug('build start');
   webpack(webpackConfig, (err, stats) => {
@@ -33,9 +36,14 @@ export default function build(opts = {}) {
 
     if (err || stats.hasErrors()) {
       if (onFail) {
-        onFail({ err, stats });
+        onFail(getErrorInfo(err, stats));
       }
-      if (!process.env.UMI_TEST) {
+
+      const isWatch = isPlainObject(webpackConfig)
+        ? webpackConfig.watch
+        : webpackConfig.some(config => config.watch); /* array */
+
+      if (!process.env.UMI_TEST && !isWatch) {
         process.exit(1);
       }
     }
@@ -57,4 +65,22 @@ export default function build(opts = {}) {
       onSuccess({ stats });
     }
   });
+}
+
+function getErrorInfo(err, stats) {
+  if (!stats.stats) {
+    return {
+      err: err || (stats.compilation && stats.compilation.errors && stats.compilation.errors[0]),
+      stats,
+      rawStats: stats,
+    };
+  }
+  const [curStats] = stats.stats;
+  return {
+    err:
+      err ||
+      (curStats.compilation && curStats.compilation.errors && curStats.compilation.errors[0]),
+    stats: curStats,
+    rawStats: stats,
+  };
 }
